Update worker.js to further improve performance (1/2)
This is the first commit out of a two-part series that improves performance further. This PR caches gate results to avoid recomputing a significant amount of data, and significantly reduces the amount of memory used by consolidating values (terms stay in the term dictionary, and we only store their identifier).